Move to external file
